fix: fix CommandTest
