+++ xmake.lua
@@ -28,6 +28,7 @@ add_requires("nlohmann_json v3.11.2")
 add_requires("rapidjson v1.1.0")
 
 -- liteldev-repo
+add_requires("dyncall 1.4")
 add_requires("compact_enc_det v1.0.1")
 add_requires("fifo_map v1.0.0")
 add_requires("pcg_cpp v1.0.0")
@@ -62,7 +63,7 @@ target("LeviLamina")
     -- xmake-repo
     add_packages("asio", "cpp-httplib", "entt", "fmt", "gsl", "gtest", "leveldb", "magic_enum", "nlohmann_json", "openssl", "rapidjson")
     -- liteldev-repo
-    add_packages("fifo_map", "pcg_cpp", "compact_enc_det", "preloader", "symbolprovider")
+    add_packages("fifo_map", "pcg_cpp", "compact_enc_det", "dyncall", "preloader", "symbolprovider")
     if has_config("localbdslibrary") then
         add_packages("localbdslibrary")
     else
